requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
catboost==1.2.5
pyarrow==15.0.2
matplotlib==3.9.0
mlflow==2.12.1
PyYAML==6.0.2
pytest==8.2.2
evidently==0.4.39